// ClaimTracker/src/generated/models/Crd2f_promotiondealsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crd2f_promotiondealsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crd2f_promotiondealsstatecode = keyof typeof Crd2f_promotiondealsstatecode;
export const Crd2f_promotiondealsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crd2f_promotiondealsstatuscode = keyof typeof Crd2f_promotiondealsstatuscode;

export interface Crd2f_promotiondealsBase {
  crd2f_brand?: string;
  crd2f_category?: string;
  "crd2f_CreatedBy@odata.bind"?: string;
  crd2f_dealname: string;
  crd2f_description?: string;
  crd2f_enddate?: string;
  crd2f_promotiondealid: string;
  crd2f_startdate?: string;
  "crd2f_Status@odata.bind"?: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crd2f_promotiondealsstatecode;
  statuscode?: Crd2f_promotiondealsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Crd2f_promotiondeals extends Crd2f_promotiondealsBase {
  crd2f_createdbyname?: string;
  crd2f_statusname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  crd2f_createdby?: object;
  _crd2f_createdby_value?: string;
  crd2f_status?: object;
  _crd2f_status_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}