// ClaimTracker/src/generated/models/Crd2f_promotionclaimsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crd2f_promotionclaimsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crd2f_promotionclaimsstatecode = keyof typeof Crd2f_promotionclaimsstatecode;
export const Crd2f_promotionclaimsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crd2f_promotionclaimsstatuscode = keyof typeof Crd2f_promotionclaimsstatuscode;

export interface Crd2f_promotionclaimsBase {
  crd2f_claimamount?: string;
  crd2f_claimdate?: string;
  crd2f_claimnumber: string;
  "crd2f_Deal@odata.bind"?: string;
  crd2f_promotionclaimid: string;
  "crd2f_ReviewedBy@odata.bind"?: string;
  "crd2f_Status@odata.bind"?: string;
  "crd2f_Vendor@odata.bind"?: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crd2f_promotionclaimsstatecode;
  statuscode?: Crd2f_promotionclaimsstatuscode;
  timezoneruleversionnumber?: string;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: string;
}

export interface Crd2f_promotionclaims extends Crd2f_promotionclaimsBase {
  crd2f_claimamount_base?: string;
  crd2f_dealname?: string;
  crd2f_reviewedbyname?: string;
  crd2f_statusname?: string;
  crd2f_vendorname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: string;
  crd2f_deal?: object;
  _crd2f_deal_value?: string;
  crd2f_reviewedby?: object;
  _crd2f_reviewedby_value?: string;
  crd2f_status?: object;
  _crd2f_status_value?: string;
  crd2f_vendor?: object;
  _crd2f_vendor_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}