// ClaimTracker/src/generated/models/Crd2f_vendorsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crd2f_vendorsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crd2f_vendorsstatecode = keyof typeof Crd2f_vendorsstatecode;
export const Crd2f_vendorsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crd2f_vendorsstatuscode = keyof typeof Crd2f_vendorsstatuscode;

export interface Crd2f_vendorsBase {
  crd2f_address?: string;
  crd2f_contactperson?: string;
  crd2f_email?: string;
  crd2f_phone?: string;
  crd2f_vendorid: string;
  crd2f_vendorname: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crd2f_vendorsstatecode;
  statuscode?: Crd2f_vendorsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Crd2f_vendors extends Crd2f_vendorsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}